increase maximal map size to 128.000 tiles (max linear width: 512)
change to the network protocol

patch #1733

[[originally from svn r17556]]

+++ common/aicore/path_finding.c
@@ -37,7 +37,7 @@
 
 /* Since speed is quite important to us and alloccation of large arrays is
  * slow, we try to pack info in the smallest types possible */
-typedef short mapindex_t;
+typedef int mapindex_t;
 typedef unsigned char utiny_t;
 
 /* ===================== Internal structures ===================== */
